Correctly handle the "nested rect" situation in the constructor
Pygame -- and consequently pgzero -- is very accommodating when constructing
a rect. You can pass four integers, two pairs of 2-tuples, or one 4-tuple.

Also, you can pass an existing Rect-like object, or an object with a .rect
attribute. The object named by the .rect attribute is either one of the above,
or it's a callable object which returns one of the above.

This is evidently a recursive solution where an object with a .rect
attribute can yield an object with a .rect attribute, and so ad infinitum.