Make Actor.rect() an undocumented method
It is better as a method because Pygame calls hasattr(actor, 'rect')
which will itself call the function and thus copy the rect; this results
in two copies. As a method there is only one copy.

It is better as undocumented because the aim is for Actor to duck-type
like a rect in its own right: it doesn't seem necessary for API
completeness to be able to copy the rect. But this could be documented
later if we decide to.