WIP #336 - Pausing, basic queue functionality
Pausing is implemented and works with both single animations and base
animations. Pausing the queue works in general but has edgecases that
cause problems.

Animation queue works but requires more work. Setting up a queue
and playing it works, complex interactions between single animations,
the queue and the base animation don't all work correctly yet.

Switched from representing not yet started animations and the queue
with -1 for the position value to representing it with None. This is
to prevent strange behaviour with unexpected checking of position
when there is no valid position. With None, the program simply errors,
exposing a location in source code to be improved.