FIX #193 - Actually working angle based movement.
Former approach was broken due to not rememberin math and python
right while not having internet...

Now, the movement functions behave properly.

One major bug is left before this can be pulled. When running
move_towards_point() every update, it results in a jittering
back and forth which is not expected until the actor overshoots
the target point. Instead it happens every single frame.